requirements.txt
streamlit
mysql-connector-python
bcrypt
plotly
fpdf
openpyxl
pillow